requirements.txt
django-crispy-forms==2.0
environs==9.5.0
pip-chill==1.0.1
pre-commit==3.2.1
asgiref==3.6.0
Django==4.1.7
marshmallow==3.19.0
packaging==23.0
python-dotenv==1.0.0
sqlparse==0.4.3
tzdata==2023.2
Pillow==9.4.0
django-recaptcha==3.0.0
django-allauth==0.53.1
django-extensions==3.2.1



